Repair geolocation service usage on iOS devices
Prior to this commit, device location was not functioning properly on
iOS devices. As of this commit, geolocation data service allows a long
enough timeout that the user may respond to iOS' default permissions
prompt such that the device is able to fetch and use device location
information normally.

This commit also makes several cleanliness/linting corrections to the
onSetUseLocation method.